Implement deployment hardening audit (A1-B9)
Backend (A1, A3, B9):
- Dockerfile bootstraps schema via entrypoint.sh before uvicorn; copies alembic.ini so alembic upgrade head runs in-image.
- WEB_CONCURRENCY drives uvicorn --workers (default 2).
- Hospital Pydantic IO schemas added; ORM model switched to Decimal for Numeric(9,6) lat/long to eliminate float roundtrip.
- auth.py coerces incoming lat/long to Decimal at the write site.

Infra (B1-B5, B7):
- Resource limits: postgres 4G/2cpu, backend 2G/1.5cpu, frontend 256M/0.5cpu.
- Log rotation via json-file max-size 10m max-file 5 on every service.
- Fail-fast interpolation with POSTGRES_PASSWORD:? and JWT_SECRET_KEY:?.
- Postgres bound to 127.0.0.1 (POSTGRES_BIND_HOST, default loopback).
- .env.prod.example expanded with secret-generation + bind-host guidance.
- HOST_PREFLIGHT.md runbook added.

AI_STATE.md appended with audit + Phase 4 Step C verification entries.

+++ apps/backend/app/models/hospital.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from sqlalchemy import Boolean, CheckConstraint, DateTime, Numeric, String, Text, func, text
@@ -27,8 +28,8 @@ class Hospital(Base):
     name: Mapped[str] = mapped_column(String(200), nullable=False, index=True)
     address: Mapped[str] = mapped_column(Text, nullable=False)
     city: Mapped[str] = mapped_column(String(100), nullable=False, index=True)
-    latitude: Mapped[float | None] = mapped_column(Numeric(9, 6), nullable=True)
-    longitude: Mapped[float | None] = mapped_column(Numeric(9, 6), nullable=True)
+    latitude: Mapped[Decimal | None] = mapped_column(Numeric(9, 6), nullable=True)
+    longitude: Mapped[Decimal | None] = mapped_column(Numeric(9, 6), nullable=True)
     contact_email: Mapped[str] = mapped_column(String(255), nullable=False)
     contact_phone: Mapped[str] = mapped_column(String(30), nullable=False)
     is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True, server_default=text("true"))
